+++ db/migrate/20221103000000_extend_tables8.rb
@@ -1,0 +1,5 @@
+class ExtendTables8 < ActiveRecord::Migration[6.0]
+  def change
+    add_column :tables, :yn_add_cloudevents_header,  :string, limit: 1, null: false, default: 'N', comment: 'Should Kafka message headers be added according to CloudEvents standard'
+  end
+end